Moving the core to .NET Standard
