+++ UserDetailsClient.Core/MainPage.xaml
@@ -1,8 +1,11 @@
-﻿<?xml version="1.0" encoding="utf-8" ?>
-<ContentPage xmlns="http://xamarin.com/schemas/2014/forms"
-             xmlns:local="clr-namespace:UserDetailsClient;assembly=UserDetailsClient"
-             xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-             x:Class="UserDetailsClient.MainPage">
+<?xml version="1.0" encoding="UTF-8"?>
+<ContentPage 
+    xmlns="http://xamarin.com/schemas/2014/forms" 
+    xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
+    xmlns:ios="clr-namespace:Xamarin.Forms.PlatformConfiguration.iOSSpecific;assembly=Xamarin.Forms.Core"
+    x:Class="UserDetailsClient.Core.MainPage"
+    Title="Azure AD B2C"
+    ios:Page.UseSafeArea="true">
     <ContentPage.Content>
         <StackLayout>
             <Label Text="MSAL Xamarin Forms Sample" VerticalOptions="Start" HorizontalTextAlignment="Center" HorizontalOptions="FillAndExpand" />
@@ -27,5 +30,4 @@
             <Button x:Name="btnSignInSignOut" Text="Sign in" Clicked="OnSignInSignOut" VerticalOptions="End" HorizontalOptions="FillAndExpand"/>
         </StackLayout>
     </ContentPage.Content>
-
 </ContentPage>
