+++ UserDetailsClient/UserDetailsClient.Droid/MainActivity.cs
@@ -1,13 +1,10 @@
-﻿using System;
-
+﻿
 using Android.App;
 using Android.Content.PM;
-using Android.Runtime;
-using Android.Views;
-using Android.Widget;
-using Android.OS;
 using Microsoft.Identity.Client;
 using Android.Content;
+
+using UserDetailsClient.Core;
 
 namespace UserDetailsClient.Droid
 {
@@ -20,7 +17,7 @@ namespace UserDetailsClient.Droid
 
             global::Xamarin.Forms.Forms.Init(this, bundle);
             LoadApplication(new App());
-            App.UiParent = new UIParent(Xamarin.Forms.Forms.Context as Activity);           
+            App.UiParent = new UIParent(Xamarin.Forms.Forms.Context as Activity);
         }
 
         protected override void OnActivityResult(int requestCode, Result resultCode, Intent data)
