+++ UserDetailsClient/UserDetailsClient.Droid/packages.config
@@ -3,7 +3,8 @@
   <package id="Microsoft.Identity.Client" version="1.1.2-preview0008" targetFramework="monoandroid81" />
   <package id="Microsoft.NETCore.Platforms" version="1.1.0" targetFramework="monoandroid60" />
   <package id="Microsoft.Win32.Primitives" version="4.3.0" targetFramework="monoandroid60" />
-  <package id="NETStandard.Library" version="1.6.1" targetFramework="monoandroid60" />
+  <package id="NETStandard.Library" version="2.0.3" targetFramework="monoandroid81" />
+  <package id="Newtonsoft.Json" version="11.0.2" targetFramework="monoandroid81" />
   <package id="System.AppContext" version="4.3.0" targetFramework="monoandroid60" />
   <package id="System.Collections" version="4.3.0" targetFramework="monoandroid60" />
   <package id="System.Collections.Concurrent" version="4.3.0" targetFramework="monoandroid60" />
@@ -20,7 +21,7 @@
   <package id="System.IO.FileSystem.Primitives" version="4.3.0" targetFramework="monoandroid60" />
   <package id="System.Linq" version="4.3.0" targetFramework="monoandroid60" />
   <package id="System.Linq.Expressions" version="4.3.0" targetFramework="monoandroid60" />
-  <package id="System.Net.Http" version="4.3.0" targetFramework="monoandroid60" />
+  <package id="System.Net.Http" version="4.3.3" targetFramework="monoandroid81" />
   <package id="System.Net.Primitives" version="4.3.0" targetFramework="monoandroid60" />
   <package id="System.Net.Sockets" version="4.3.0" targetFramework="monoandroid60" />
   <package id="System.ObjectModel" version="4.3.0" targetFramework="monoandroid60" />
@@ -46,7 +47,7 @@
   <package id="System.Threading" version="4.3.0" targetFramework="monoandroid60" />
   <package id="System.Threading.Tasks" version="4.3.0" targetFramework="monoandroid60" />
   <package id="System.Threading.Timer" version="4.3.0" targetFramework="monoandroid60" />
-  <package id="System.Xml.ReaderWriter" version="4.3.0" targetFramework="monoandroid60" />
+  <package id="System.Xml.ReaderWriter" version="4.3.1" targetFramework="monoandroid81" />
   <package id="System.Xml.XDocument" version="4.3.0" targetFramework="monoandroid60" />
   <package id="Xamarin.Android.Support.Animated.Vector.Drawable" version="25.4.0.2" targetFramework="monoandroid81" />
   <package id="Xamarin.Android.Support.Annotations" version="25.4.0.2" targetFramework="monoandroid81" />
@@ -65,5 +66,5 @@
   <package id="Xamarin.Android.Support.v7.Palette" version="25.4.0.2" targetFramework="monoandroid81" />
   <package id="Xamarin.Android.Support.v7.RecyclerView" version="25.4.0.2" targetFramework="monoandroid81" />
   <package id="Xamarin.Android.Support.Vector.Drawable" version="25.4.0.2" targetFramework="monoandroid81" />
-  <package id="Xamarin.Forms" version="2.5.0.91635" targetFramework="monoandroid81" />
+  <package id="Xamarin.Forms" version="3.0.0.482510" targetFramework="monoandroid81" />
 </packages>
