+++ UserDetailsClient/UserDetailsClient.iOS/packages.config
@@ -1,5 +1,6 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
   <package id="Microsoft.Identity.Client" version="1.1.2-preview0008" targetFramework="xamarinios10" />
-  <package id="Xamarin.Forms" version="2.5.0.91635" targetFramework="xamarinios10" />
+  <package id="Newtonsoft.Json" version="11.0.2" targetFramework="xamarinios10" />
+  <package id="Xamarin.Forms" version="3.0.0.482510" targetFramework="xamarinios10" />
 </packages>
