--- UserDetailsClient/UserDetailsClient/App.cs
@@ -1,58 +1,0 @@
-﻿using Microsoft.Identity.Client;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-
-using Xamarin.Forms;
-
-namespace UserDetailsClient
-{
-    public class App : Application
-    {
-        public static PublicClientApplication PCA = null;
-        
-        // Azure AD B2C Coordinates
-        public static string Tenant = "fabrikamb2c.onmicrosoft.com";
-        public static string ClientID = "90c0fe63-bcf2-44d5-8fb7-b8bbc0b29dc6";
-        public static string PolicySignUpSignIn = "b2c_1_susi";
-        public static string PolicyEditProfile = "b2c_1_edit_profile";
-        public static string PolicyResetPassword = "b2c_1_reset";
-
-        public static string[] Scopes = { "https://fabrikamb2c.onmicrosoft.com/demoapi/demo.read" };
-        public static string ApiEndpoint = "https://fabrikamb2chello.azurewebsites.net/hello";
-
-        public static string AuthorityBase = $"https://login.microsoftonline.com/tfp/{Tenant}/";
-        public static string Authority = $"{AuthorityBase}{PolicySignUpSignIn}";
-        public static string AuthorityEditProfile = $"{AuthorityBase}{PolicyEditProfile}";
-        public static string AuthorityPasswordReset = $"{AuthorityBase}{PolicyResetPassword}";
-
-        public static UIParent UiParent = null;
-
-      
-
-        public App()
-        {
-            // default redirectURI; each platform specific project will have to override it with its own
-            PCA = new PublicClientApplication(ClientID, Authority);
-            PCA.RedirectUri = $"msal{ClientID}://auth";
-                        
-            MainPage = new NavigationPage(new MainPage());        
-        }
-
-        protected override void OnStart()
-        {
-            // Handle when your app starts
-        }
-
-        protected override void OnSleep()
-        {
-            // Handle when your app sleeps
-        }
-
-        protected override void OnResume()
-        {
-            // Handle when your app resumes
-        }
-    }
-}
