--- UserDetailsClient/UserDetailsClient/GettingStarted.Xamarin
@@ -1,4 +1,0 @@
-<GettingStarted>
-	<LocalContent>GS\XF\CS\App\GettingStarted.html</LocalContent>
-	<EmbeddedNavigation>false</EmbeddedNavigation>
-</GettingStarted>
