--- UserDetailsClient/UserDetailsClient/packages.config
@@ -1,41 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="Microsoft.Identity.Client" version="1.1.2-preview0008" targetFramework="portable45-net45+win8" />
-  <package id="Microsoft.NETCore.Platforms" version="1.1.0" targetFramework="portable45-net45+win8" />
-  <package id="Microsoft.NETCore.Targets" version="1.1.0" targetFramework="portable45-net45+win8" />
-  <package id="NETStandard.Library" version="1.6.1" targetFramework="portable45-net45+win8" />
-  <package id="Newtonsoft.Json" version="8.0.3" targetFramework="portable45-net45+win8" />
-  <package id="runtime.native.System" version="4.3.0" targetFramework="portable45-net45+win8" />
-  <package id="System.Collections" version="4.3.0" targetFramework="portable45-net45+win8" />
-  <package id="System.Collections.Concurrent" version="4.3.0" targetFramework="portable45-net45+win8" />
-  <package id="System.Diagnostics.Debug" version="4.3.0" targetFramework="portable45-net45+win8" />
-  <package id="System.Diagnostics.Tools" version="4.3.0" targetFramework="portable45-net45+win8" />
-  <package id="System.Diagnostics.Tracing" version="4.3.0" targetFramework="portable45-net45+win8" />
-  <package id="System.Globalization" version="4.3.0" targetFramework="portable45-net45+win8" />
-  <package id="System.IO" version="4.3.0" targetFramework="portable45-net45+win8" />
-  <package id="System.IO.Compression" version="4.3.0" targetFramework="portable45-net45+win8" />
-  <package id="System.Linq" version="4.3.0" targetFramework="portable45-net45+win8" />
-  <package id="System.Linq.Expressions" version="4.3.0" targetFramework="portable45-net45+win8" />
-  <package id="System.Net.Http" version="4.3.0" targetFramework="portable45-net45+win8" />
-  <package id="System.Net.Primitives" version="4.3.0" targetFramework="portable45-net45+win8" />
-  <package id="System.ObjectModel" version="4.3.0" targetFramework="portable45-net45+win8" />
-  <package id="System.Reflection" version="4.3.0" targetFramework="portable45-net45+win8" />
-  <package id="System.Reflection.Extensions" version="4.3.0" targetFramework="portable45-net45+win8" />
-  <package id="System.Reflection.Primitives" version="4.3.0" targetFramework="portable45-net45+win8" />
-  <package id="System.Resources.ResourceManager" version="4.3.0" targetFramework="portable45-net45+win8" />
-  <package id="System.Runtime" version="4.3.0" targetFramework="portable45-net45+win8" />
-  <package id="System.Runtime.Extensions" version="4.3.0" targetFramework="portable45-net45+win8" />
-  <package id="System.Runtime.InteropServices" version="4.3.0" targetFramework="portable45-net45+win8" />
-  <package id="System.Runtime.InteropServices.RuntimeInformation" version="4.3.0" targetFramework="portable45-net45+win8" />
-  <package id="System.Runtime.Numerics" version="4.3.0" targetFramework="portable45-net45+win8" />
-  <package id="System.Runtime.Serialization.Json" version="4.3.0" targetFramework="portable45-net45+win8" />
-  <package id="System.Runtime.Serialization.Primitives" version="4.3.0" targetFramework="portable45-net45+win8" />
-  <package id="System.Text.Encoding" version="4.3.0" targetFramework="portable45-net45+win8" />
-  <package id="System.Text.Encoding.Extensions" version="4.3.0" targetFramework="portable45-net45+win8" />
-  <package id="System.Text.RegularExpressions" version="4.3.0" targetFramework="portable45-net45+win8" />
-  <package id="System.Threading" version="4.3.0" targetFramework="portable45-net45+win8" />
-  <package id="System.Threading.Tasks" version="4.3.0" targetFramework="portable45-net45+win8" />
-  <package id="System.Xml.ReaderWriter" version="4.3.0" targetFramework="portable45-net45+win8" />
-  <package id="System.Xml.XDocument" version="4.3.0" targetFramework="portable45-net45+win8" />
-  <package id="Xamarin.Forms" version="2.5.0.91635" targetFramework="portable45-net45+win8" />
-</packages>
